# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexWatch Hub API
  description: >
    Custom API routes exposed by the NexWatch hub (a PocketBase application)
    under /api/custom/*, plus the unauthenticated /healthz liveness probe.
    PocketBase's own built-in routes (record CRUD under /api/collections/*,
    auth, realtime, files, etc.) are not documented here — see the
    PocketBase API reference for those.
  version: "1.0.0"
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/public/status:
    get:
      operationId: getPublicStatus
      summary: Public status page data (unauthenticated)
      description: >
        Returns the allowlisted status-page items (checks and/or agents,
        each with an admin-typed label) configured via the
        status_page_items setting, with their live status, uptime
        percentages, and daily history. Never exposes a check's target,
        an agent's hostname/IP, or either's raw id — only the label an
        admin explicitly typed into status_page_items. Returns 404 when
        status_page_enabled is false. Unauthenticated, rate-limited to 60
        requests/minute per client IP, and cacheable for 30 seconds.
      security: []
      tags: [status-page]
      responses:
        "200":
          description: Status page data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  title: { type: string }
                  description: { type: string }
                  overall: { type: string, enum: [operational, degraded, down, unknown] }
                  updated_at: { type: string, format: date-time }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        type: { type: string, enum: [check, agent] }
                        label: { type: string }
                        status: { type: string, enum: [operational, degraded, down, unknown] }
                        uptime_24h: { type: number }
                        uptime_7d: { type: number }
                        uptime_30d: { type: number }
                        latency_ms: { type: number, description: "checks only." }
                        daily:
                          type: array
                          items:
                            type: object
                            properties:
                              date: { type: string, format: date }
                              uptime: { type: number }
                              incidents: { type: integer }
                        last_incident_at: { type: string, format: date-time }
        "404":
          description: The status page is not enabled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A PocketBase auth token (superuser or "users" record) passed in the
        Authorization header. PocketBase accepts it either as a bare token
        or with a "Bearer " prefix.
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...